% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casplib.R
\name{casp.agg.linexloss}
\alias{casp.agg.linexloss}
\title{Coordinate-wise shrinkage prediction for aggregated targets under Linex loss}
\usage{
casp.agg.linexloss(X, S, A, a, tau, beta, eta, mx, mw, m0, type)
}
\arguments{
\item{X}{a \eqn{mx\times n} matrix of past observations.}

\item{S}{\eqn{n\times n} sample covariance matrix based on \eqn{mw} samples}

\item{A}{a \eqn{p\times n} aggregation matrix with \eqn{rank(\mathbf{A}\mathbf{A}^T)=p}.}

\item{a}{this is a \eqn{p\times 1} vector of Linex loss parameter. See equation (4)
in the casp paper.}

\item{tau}{the positive scale hyper-parameter for the prior on the locations}

\item{beta}{the non-negative shape hyper-parameter for the prior on the locations}

\item{eta}{the prior mean of the locations}

\item{mx}{the sample size of past observations \eqn{\mathbf{X}}}

\item{mw}{the size of the side information for calculating the sample covariance
matrix \eqn{\mathbf{S}}}

\item{m0}{the sample size of the future observation. Usually this is set to 1}

\item{type}{if type = 1 then all the shrinkage factors are equal to 1. If type = 0, then
the shrinkage factors are estimated.}
}
\value{
\enumerate{
\item q - aggregated shrinkage prediction under Linex loss for the \eqn{p} coordinates
\item f - estimated shrinkage factors (equal to \eqn{\mathbf{1}} if type = 1)
}
}
\description{
Main function for aggregated shrinkage prediction in high-dimensional, non-exchangeable hierarchical Gaussian models
with an unknown location as well as an unknown spiked covariance structure under Linex loss.
}
\details{
This function is based on Definition 4 of the casp paper, and relies on
\code{\link{rmt.est}} and \code{\link{Gfun.est}}. The shrinkage factors
are estimated using the formulation given in Definition 4 and use
\code{\link{f.agg.est}} in the background. Please see the casp paper in the reference
for more details about these estimation techniques. If \eqn{(\tau,\beta)} are unknown
then one may first use \code{\link{taubeta.casp.est}} to estimate them and then
use the estimated values in \code{\link{casp.agg.checkloss}}.
}
\examples{
library(casp)
n = 10
p = 3
set.seed(42)
A = matrix(runif(n*p,0,1),p,n)
mx = 5
S = diag(c(10,8,6,4,rep(1,n-4)))
set.seed(42)
X<- matrix(runif(mx*n),mx,n)
tau = 1
beta = 0.5
eta = rep(0,n)
mw = 100
m0 = 1
a = rep(-1,p)
q.casp<- casp.agg.linexloss(X,S,A,a,tau,beta,eta,mx,mw,m0,0)

}
\references{
\enumerate{
\item Trambak Banerjee, Gourab Mukherjee, and Debashis Paul. Improved Shrinkage Prediction under a Spiked
Covariance Structure, 2021.
}
}
\seealso{
\code{\link{casp.agg.linexloss}},\code{\link{f.agg.est}},\code{\link{rmt.est}},\code{\link{taubeta.casp.est}}
}
