% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casplib.R
\name{eigen.est}
\alias{eigen.est}
\title{eigen.est}
\usage{
eigen.est(l, l.0, rho, K, n)
}
\arguments{
\item{l}{eigenvalues of the sample covariance matrix}

\item{l.0}{an estimate of noise via maximum likelihood}

\item{rho}{the ratio \eqn{n/(mw-1)} where \eqn{mw} is the sample size}

\item{K}{the number of spikes}

\item{n}{the dimension of the covariance matrix}
}
\value{
\enumerate{
\item l0.hat - bias corrected estimate of the unknown noise level
\item l.hat - bias corrected estimates of the leading K eignevalues of \eqn{\mathbf{\Sigma}}
}
}
\description{
Provides efficient and bias corrected estimates of the leading eigenvalues
of \eqn{\mathbf{\Sigma}} under a spiked covariance model.
}
\details{
This function is called by \code{\link{rmt.est}} and the estimate of noise, l.0, that
is used here is \eqn{\ell_0=(n-K)^{-1}\sum_{j=K+1}^{n}\ell_j}.
}
\examples{
library(casp)
K = 4
n = 10
l = c(10,8,6,4,rep(1,6))
l.0 = ((n-K)^{-1})*sum(l[(K+1):n])
mw = 100
rho = n/(mw-1)
eigen.est(l,l.0,rho,K,n)

}
\references{
\enumerate{
\item Debashis Paul. Asymptotics of sample eigenstructure for a large dimensional spiked covariance
model. Statistica Sinica, pages 1617-1642, 2007.
\item Alexei Onatski. Asymptotics of the principal components estimator of large factor models
with weakly influential factors. Journal of Econometrics, 168(2):244-258, 2012.
\item Damien Passemier, Zhaoyuan Li, and Jianfeng Yao. On estimation of the noise variance
in high dimensional probabilistic principal component analysis. Journal of the Royal
Statistical Society: Series B (Statistical Methodology), 79(1):51-67, 2017.
}
}
\seealso{
\code{\link{rmt.est}}, \code{\link{casp.checkloss}}, \code{\link{casp.linexloss}}
}
