% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casplib.R
\name{f.agg.est}
\alias{f.agg.est}
\title{Estimates the coordinate-wise shrinkage factors for aggregated prediction}
\usage{
f.agg.est(rmt, A, beta, tau, mx)
}
\arguments{
\item{rmt}{the output from \code{\link{rmt.est}}}

\item{A}{a \eqn{p\times n} aggregation matrix with \eqn{rank(\mathbf{A}\mathbf{A}^T)=p}.}

\item{beta}{the non-negative shape hyper-parameter for the prior on the locations}

\item{tau}{the positive scale hyper-parameter for the prior on the locations}

\item{mx}{the sample size of the past observations \eqn{\mathbf{X}}}
}
\value{
\enumerate{
\item f - estimated shrinkage factors
}
}
\description{
Estimates the coordinate-wise shrinkage factors that are involved in aggregated prediction
using the information in the sample covariance matrix \eqn{\mathbf{S}}. See Definition (4) of
the casp paper in the reference.
}
\details{
This function relies on the output from \code{\link{rmt.est}} and is called by \code{\link{casp.agg.checkloss}},
and \code{\link{casp.agg.linexloss}}. Please see Definition (4) of the casp paper in the reference
for more details.
}
\examples{
library(casp)
tau = 1
beta = 0.5
mx = 1
K = 4
S = diag(c(10,8,6,4,rep(1,6)))
mw = 50
rmt<- rmt.est(K,S,mw)
f<- f.est(rmt,beta,tau,mx)

}
\references{
\enumerate{
\item Trambak Banerjee, Gourab Mukherjee, and Debashis Paul. Improved Shrinkage Prediction under a Spiked
Covariance Structure, 2021.
}
}
\seealso{
\code{\link{rmt.est}}, \code{\link{casp.agg.checkloss}}, \code{\link{casp.agg.linexloss}}
}
