% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casplib.R
\name{f.est}
\alias{f.est}
\title{f.est}
\usage{
f.est(rmt, beta, tau, mx)
}
\arguments{
\item{rmt}{the output from \code{\link{rmt.est}}}

\item{beta}{the non-negative shape hyper-parameter for the prior on the locations}

\item{tau}{the positive scale hyper-parameter for the prior on the locations}

\item{mx}{the sample size of the past observations \eqn{\mathbf{X}}}
}
\value{
\enumerate{
\item f - estimated shrinkage factors
}
}
\description{
Estimates the coordinate-wise shrinkage factors using the information in
the sample covariance matrix \eqn{\mathbf{S}}. See Definition (3) of
the casp paper (see the reference).
}
\details{
This function relies on the output from \code{\link{rmt.est}} and is called by \code{\link{casp.checkloss}},
and \code{\link{casp.linexloss}}. Please see Definition (3) of the casp paper in the reference
for more details.
}
\examples{
library(casp)
tau = 1
beta = 0.5
mx = 1
K = 4
S = diag(c(10,8,6,4,rep(1,6)))
mw = 50
rmt<- rmt.est(K,S,mw)
f<- f.est(rmt,beta,tau,mx)

}
\references{
\enumerate{
\item Trambak Banerjee, Gourab Mukherjee, and Debashis Paul. Improved Shrinkage Prediction under a Spiked
Covariance Structure, 2021.
}
}
\seealso{
\code{\link{rmt.est}}, \code{\link{casp.checkloss}}, \code{\link{casp.linexloss}}
}
