% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casplib.R
\name{g.est}
\alias{g.est}
\title{g.est}
\usage{
g.est(rmt, beta, tau, mx)
}
\arguments{
\item{rmt}{the output from \code{\link{rmt.est}}.}

\item{beta}{the non-negative shape hyper-parameter for the prior on the locations.}

\item{tau}{the positive scale hyper-parameter for the prior on the locations.}

\item{mx}{the sample size of the past observations \eqn{\mathbf{X}}.}
}
\value{
\enumerate{
\item G - an estimate of \eqn{\tau H_{-1,1+\beta,\beta}}
\item Ginv - an estimate of \eqn{\tau^{-1} H_{1,-1-\beta,\beta}}
}
}
\description{
Provides an estimate of (1) \eqn{\tau H_{-1,1+\beta,\beta}}, and
(2) \eqn{\tau^{-1} H_{1,-1-\beta,\beta}} using equation (13) of
the casp paper (see the reference).
}
\details{
This function relies on the output from \code{\link{rmt.est}} and is called by \code{\link{casp.checkloss}},
and \code{\link{taubeta.casp.est}}. Please see equation (13) of the casp paper in the reference
to get more details about the estimation.
}
\examples{
library(casp)
tau = 1
beta = 0.5
mx = 1
K = 4
S = diag(c(10,8,6,4,rep(1,6)))
mw = 50
rmt<- rmt.est(K,S,mw)
g.out<- g.est(rmt,beta,tau,mx)

}
\references{
\enumerate{
\item Trambak Banerjee, Gourab Mukherjee, and Debashis Paul. Improved Shrinkage Prediction under a Spiked
Covariance Structure, 2021.
}
}
\seealso{
\code{\link{rmt.est}}, \code{\link{taubeta.casp.est}}
}
