% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casplib.R
\name{Gfun.est}
\alias{Gfun.est}
\title{Provides an estimate of the function \eqn{\mathbf{G}_{r,\alpha,\beta}(\mathbf{\Sigma},\mathbf{A})}}
\usage{
Gfun.est(A, r, alpha, beta, tau, rmt, mx)
}
\arguments{
\item{A}{a \eqn{p\times n} aggregation matrix with \eqn{rank(\mathbf{A}\mathbf{A}^T)=p}}

\item{r}{a value in \eqn{\{-1,0,1\}}}

\item{alpha}{a non-negative number}

\item{beta}{the non-negative shape hyper-parameter for the prior on the locations}

\item{tau}{the positive scale hyper-parameter for the prior on the locations}

\item{rmt}{the output from \code{\link{rmt.est}}}

\item{mx}{the sample size of the past observations \eqn{\mathbf{X}}}
}
\value{
\enumerate{
\item Gfun - an estimate of \eqn{G_{r,\alpha,\beta}(\mathbf{\Sigma},\mathbf{A})}
}
}
\description{
Provides an estimate of the function
\eqn{\mathbf{G}_{r,\alpha,\beta}(\mathbf{\Sigma},\mathbf{A})} using the display in
Section 4 of the casp paper in the reference.
}
\details{
This function relies on the output from \code{\link{rmt.est}} and is called by \code{\link{casp.agg.checkloss}}, and
\code{\link{casp.agg.linexloss}}. Please see Section 4 of the casp
paper in the reference to get more details about the estimation.
}
\examples{
library(casp)
n = 10
p = 3
set.seed(42)
A = matrix(runif(n*p,0,1),p,n)
r = 1
alpha = 0
tau = 1
beta = 0.5
mx = 1
K = 4
S = diag(c(10,8,6,4,rep(1,6)))
mw = 50
rmt<- rmt.est(K,S,mw)
Gfun.out<- Gfun.est(A,r,alpha,beta,tau,rmt,mx)

}
\references{
\enumerate{
\item Trambak Banerjee, Gourab Mukherjee, and Debashis Paul. Improved Shrinkage Prediction under a Spiked
Covariance Structure, 2021.
}
}
\seealso{
\code{\link{rmt.est}}, \code{\link{casp.agg.checkloss}}, \code{\link{casp.agg.linexloss}}
}
