% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casplib.R
\name{hfun.est}
\alias{hfun.est}
\title{hfun.est}
\usage{
hfun.est(r, alpha, beta, rmt, mx, tau)
}
\arguments{
\item{r}{a value in \eqn{\{-1,0,1\}}}

\item{alpha}{a non-negative number}

\item{beta}{the non-negative shape hyper-parameter for the prior on the locations}

\item{rmt}{the output from \code{\link{rmt.est}}}

\item{mx}{the sample size of the past observations \eqn{\mathbf{X}}}

\item{tau}{the positive scale hyper-parameter for the prior on the locations}
}
\value{
\enumerate{
\item h - an estimate of \eqn{H_{r,\alpha,\beta}}
}
}
\description{
Provides an estimate of \eqn{\mathbf{H}_{r,\alpha,\beta}} using equation (13) of
the casp paper (see the reference).
}
\details{
This function relies on the output from \code{\link{rmt.est}} and is called by \code{\link{casp.checkloss}},
and \code{\link{casp.linexloss}}. Please see equation (13) of the casp paper in the reference
to get more details about the estimation.
}
\examples{
library(casp)
r = -1
alpha = 0
tau = 1
beta = 0.5
mx = 1
K = 4
S = diag(c(10,8,6,4,rep(1,6)))
mw = 50
rmt<- rmt.est(K,S,mw)
H.hat<- hfun.est(r,alpha,beta,rmt,mx,tau)

}
\references{
\enumerate{
\item Trambak Banerjee, Gourab Mukherjee, and Debashis Paul. Improved Shrinkage Prediction under a Spiked
Covariance Structure, 2021.
}
}
\seealso{
\code{\link{rmt.est}}, \code{\link{casp.checkloss}}, \code{\link{casp.linexloss}}
}
