% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikelib.R
\name{KN.test}
\alias{KN.test}
\title{KN.test}
\usage{
KN.test(S, numObs, alpha = 0.05)
}
\arguments{
\item{S}{the sample covariance matrix}

\item{numObs}{number of data observations}

\item{alpha}{significance level for the test}
}
\value{
\enumerate{
\item numOfSpikes - The number of spikes K
\item sigma.sq - Estimated noise variance
}
}
\description{
Estimate the number of spikes in a spiked covariance model using the procedure described in Kritchman and Nadler, (2009)
}
\details{
This method calculates the number of spikes in a spiked covariance model
by iteratively comparing a standardized eigenvalue to the quantile of Tracy-Wisdom
distribution for given significance level. The procedure is described in Kritchman and Nadler, (2009).
}
\examples{
library(casp)
S<-diag(c(10,8,6,4,rep(1,6)))
KN.test.out<-KN.test(S,50)

}
\references{
Kritchman, S. and Nadler, B., 2009. Non-parametric detection of the number of signals: Hypothesis testing and random matrix theory.
IEEE Transactions on Signal Processing, 57(10), pp.3930-3941.
}
\seealso{
\code{\link{sigmasq.est}}
}
