% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casplib.R
\name{rmt.est}
\alias{rmt.est}
\title{Estimation of leading eigenvalues and adjustment factors for the sample eigenvectors}
\usage{
rmt.est(K, S, mw)
}
\arguments{
\item{K}{the number of spikes}

\item{S}{the sample covariance matrix}

\item{mw}{sample size}
}
\value{
\enumerate{
\item l0.hat - bias corrected estimate of the unknown noise level
\item l.hat - bias corrected estimates of the leading K eignevalues of Sigma
\item pj - sample eigenvectors of \eqn{\mathbf{S}}
\item zeta - asymptotic adjustment factors for the eigenvectors of S
\item K - the number of spikes (provided as an input)
}
}
\description{
Provides (1) efficient and bias corrected estimates of the leading eigenvalues
of \eqn{\mathbf{Sigma}} under a spiked covariance model, and (2) asymptotic adjustment factors
for the sample eigenvectors.
}
\details{
This function is called by \code{\link{casp.checkloss}}, \code{\link{casp.linexloss}}
and their counterparts for aggregated prediction.
}
\examples{
library(casp)
K = 4
S = diag(c(10,8,6,4,rep(1,6)))
mw = 50
rmt.out<- rmt.est(K,S,mw)

}
\references{
\enumerate{
\item Debashis Paul. Asymptotics of sample eigenstructure for a large dimensional spiked covariance
model. Statistica Sinica, pages 1617-1642, 2007.
\item Alexei Onatski. Asymptotics of the principal components estimator of large factor models
with weakly influential factors. Journal of Econometrics, 168(2):244-258, 2012.
\item Damien Passemier, Zhaoyuan Li, and Jianfeng Yao. On estimation of the noise variance
in high dimensional probabilistic principal component analysis. Journal of the Royal
Statistical Society: Series B (Statistical Methodology), 79(1):51-67, 2017.
}
}
\seealso{
\code{\link{eigen.est}}, \code{\link{casp.checkloss}}, \code{\link{casp.linexloss}},
\code{\link{casp.agg.checkloss}}, \code{\link{casp.agg.linexloss}}
}
