% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikelib.R
\name{sigmasq.est}
\alias{sigmasq.est}
\title{Estimation of the unknown noise variance}
\usage{
sigmasq.est(lambdas, numObs, numOfSpikes)
}
\arguments{
\item{lambdas}{eigenvalues to which the distribution is fitted}

\item{numObs}{number of data observations}

\item{numOfSpikes}{number of spikes in the spike covariance model}
}
\value{
\enumerate{
\item sigma.sq - Estimated noise variance
}
}
\description{
Estimates the unknown noise variance based on the procedure described in Kritchman and Nadler, (2009)
}
\details{
This method calculates an unbiased estimate of the noise variance
based on the procedure described in Kritchman and Nadler, (2009). It involves
solving a couple of nonlinear equations. This function is called by \code{\link{KN.test}}.
}
\examples{
library(casp)
lambdas<-c(10,8,6,4,rep(1,6))
sigmasq.est(lambdas,100,4)

}
\references{
Kritchman, S. and Nadler, B., 2009. Non-parametric detection of the number of signals: Hypothesis testing and random matrix theory.
IEEE Transactions on Signal Processing, 57(10), pp.3930-3941.
}
\seealso{
\code{\link{KN.test}}
}
