% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casplib.R
\name{taubeta.casp.est}
\alias{taubeta.casp.est}
\title{Data-driven estimation of the hyper-parameters \eqn{(\tau,\beta)}}
\usage{
taubeta.casp.est(grid.val, rmt, mx, X)
}
\arguments{
\item{grid.val}{a matrix with two columns. Each row of the matrix represents an
element of a two dimensional grid. The first component represents a likely value for \eqn{\tau}
while the second component is a likely value for \eqn{\beta}.}

\item{rmt}{the output from \code{\link{rmt.est}}.}

\item{mx}{the sample size of the past observations \eqn{\mathbf{X}}.}

\item{X}{a \eqn{mx\times n} matrix of past observations.}
}
\value{
\enumerate{
\item An estimate of \eqn{(\tau,\beta)}
}
}
\description{
Provides an estimate of the hyper-parameters \eqn{(\tau,\beta)} using equation (15) of
the casp paper in the reference.
}
\details{
This function relies on the output from \code{\link{rmt.est}} and calls \code{\link{g.est}}.
Please see Section 3.3 of the casp paper for more details.
}
\examples{
library(casp)
set.seed(42)
n = 10
mx = 5
X = matrix(runif(mx*n),mx,n)
K = 4
S = diag(c(10,8,6,4,rep(1,n-4)))
mw = 50
rmt<- rmt.est(K,S,mw)
tau.grid<-c(0.2,0.3,0.4,0.5)
beta.grid<-c(0.15,0.25,0.35,0.5)
grid.val<- cbind(rep(tau.grid,each = length(beta.grid)),
                    rep(beta.grid,length(beta.grid)))
taubeta.estimated<- taubeta.casp.est(grid.val,rmt,mx,X)

}
\references{
\enumerate{
\item Trambak Banerjee, Gourab Mukherjee, and Debashis Paul. Improved Shrinkage Prediction under a Spiked
Covariance Structure, 2021.
}
}
\seealso{
\code{\link{rmt.est}}, \code{\link{casp.checkloss}}, \code{\link{casp.linexloss}}
}
